@@ -9,12 +9,23 @@
     "start": "vue-cli-service serve"
   },
   "dependencies": {
-    "axios": "^1.7.2",
+    "axios": "^1.7.3",
     "core-js": "^3.8.3",
+    "jquery": "^3.7.1",
+    "js-cookie": "^3.0.5",
+    "jwt-decode": "^4.0.0",
     "pinia": "^2.2.0",
     "pinia-plugin-persistedstate": "^3.2.1",
-    "vue": "^3.2.13",
-    "vue-router": "^4.4.1"
+    "vee-validate": "^4.13.2",
+    "vue": "^3.4.35",
+    "vue-plugin-load-script": "^2.1.1",
+    "vue-router": "^4.4.1",
+    "vue-slick-carousel": "^1.0.6",
+    "vue3-carousel": "^0.3.3",
+    "yarn": "^1.22.22",
+    "yup": "^1.4.0",
+    "webstomp-client": "^1.2.6",
+    "sockjs-client": "^1.6.1"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
