@@ -26,10 +26,10 @@
           </div>
         </div>
         <div class="participations">
+          <div>유송연</div>
           <div>서시현</div>
           <div>최승은</div>
           <div>곽동현</div>
-          <div>유송연</div>
           <div>최수빈</div>
         </div>
       </div>
@@ -42,7 +42,6 @@
 
 <style scoped>
 footer{
-  margin-top: 5rem;
   padding: 1.75rem 0;
   border-color: hsla(220, 9%, 46%, .3);
   border-top-width: 1px;
