@@ -1,9 +1,26 @@
 <script setup>
-
 import MainHeader from "@/components/layout/MainHeader.vue";
-import PostList from "@/pages/component/List/PostList.vue";
+import PostList from "@/components/post/List/PostList.vue";
 import MainFooter from "@/components/layout/MainFooter.vue";
-// import SearchBar from "@/pages/component/Menu/SearchBar.vue";
+
+import { useFreePostStore } from "@/pages/Community/FreeBoard/stores/useFreePostStore";
+import { useOpenPostStore } from "@/pages/Community/OpenBoard/stores/useOpenPostStore";
+import {onMounted, ref} from "vue";
+
+const freePostStore = useFreePostStore();
+const openPostStore = useOpenPostStore();
+const isDataLoaded = ref(true);
+
+onMounted( async ()=> {
+  try {
+    await freePostStore.readAllPosts(1, 10);
+    await openPostStore.readAllPosts(1, 10);
+    isDataLoaded.value = true;
+  } catch (error) {
+    console.error("Error loading posts:", error);
+  }
+})
+
 </script>
 
 <template>
@@ -11,8 +28,8 @@ import MainFooter from "@/components/layout/MainFooter.vue";
     <MainHeader></MainHeader>
     <main>
       <div class="content-area">
-        <PostList title="자유 게시판"></PostList>
-        <PostList title="공개 게시판"></PostList>
+        <PostList :data-list="freePostStore.posts" title="자유 게시판" board="free"></PostList>
+        <PostList :data-list="openPostStore.posts" title="공개 게시판" board="open"></PostList>
       </div>
     </main>
     <MainFooter></MainFooter>
@@ -20,6 +37,9 @@ import MainFooter from "@/components/layout/MainFooter.vue";
 </template>
 
 <style scoped>
+main{
+  align-items: flex-start !important;
+}
 .content-area{
   display: flex;
   width: 100%;
