@@ -1,10 +1,17 @@
 <script setup>
 
+import MainHeader from "@/components/layout/MainHeader.vue";
+import PostEditor from "@/components/post/Detail/PostEditor.vue";
+import MainFooter from "@/components/layout/MainFooter.vue";
 </script>
 
 <template>
-  <div>
-    <h1>공지사항 수정 페이지</h1>
+  <div class="body-container">
+    <MainHeader></MainHeader>
+    <main>
+      <PostEditor board-title="공지사항 수정"></PostEditor>
+    </main>
+    <MainFooter></MainFooter>
   </div>
 </template>
 
