@@ -1,8 +1,6 @@
+
 <template>
   <div class="container-wrap">
-    <div class="title">
-      <h1>좌석 예약</h1>
-    </div>
     <div class="container">
       <div class="seat-wrap">
         <div class="seat-table-wrap f3">
@@ -13,21 +11,27 @@
         <div class="seat-table-wrap f4">
           <p class="seat-table-label">4층</p>
           <button class="seat-table" :class="{ 'selected': selectedSection === 'A' && selectedFloor === 4 }"
-                  @click="getTime(4, 'A')">A</button>
+                  @click="getTime(4, 'A')">A
+          </button>
         </div>
         <div class="seat-table-wrap f5">
           <p class="seat-table-label">5층</p>
           <div class="seat-table-fifth">
             <button class="seat-table table-1" :class="{ 'selected': selectedSection === 'A' && selectedFloor === 5 }"
-                    @click="getTime(5, 'A')">A</button>
+                    @click="getTime(5, 'A')">A
+            </button>
             <button class="seat-table" :class="{ 'selected': selectedSection === 'B' && selectedFloor === 5 }"
-                    @click="getTime(5, 'B')">B</button>
+                    @click="getTime(5, 'B')">B
+            </button>
             <button class="seat-table t3" :class="{ 'selected': selectedSection === 'C' && selectedFloor === 5 }"
-                    @click="getTime(5, 'C')">C</button>
+                    @click="getTime(5, 'C')">C
+            </button>
             <button class="seat-table t4" :class="{ 'selected': selectedSection === 'D' && selectedFloor === 5 }"
-                    @click="getTime(5, 'D')">D</button>
+                    @click="getTime(5, 'D')">D
+            </button>
             <button class="seat-table t5" :class="{ 'selected': selectedSection === 'E' && selectedFloor === 5 }"
-                    @click="getTime(5, 'E')">E</button>
+                    @click="getTime(5, 'E')">E
+            </button>
           </div>
         </div>
         <div class="seating-status-wrap">
@@ -51,7 +55,8 @@
             <div class="time-select-wrap">
               <p>{{ selectedFloor }}층 {{ selectedSection }}테이블 예약 가능 시간</p>
 
-              <div class="time-selector" v-if="reservationStore.reservationTimeList && reservationStore.reservationTimeList.seatIdx && reservationStore.reservationTimeList.seatIdx.length > 0">
+              <div class="time-selector"
+                   v-if="reservationStore.reservationTimeList && reservationStore.reservationTimeList.seatIdx && reservationStore.reservationTimeList.seatIdx.length > 0">
                 <div v-for="(seat, index) in reservationStore.reservationTimeList.seatIdx" :key="index">
                   <input
                       type="checkbox"
@@ -74,8 +79,8 @@
 </template>
 
 <script>
-import { mapStores } from 'pinia';
-import { useReservationStore } from '../stores/useReservationStore';
+import {mapStores} from 'pinia';
+import {useReservationStore} from '../stores/useReservationStore';
 
 export default {
   name: 'ReservationComponent',
@@ -123,16 +128,17 @@ export default {
 </script>
 
 <style>
-.container-wrap{
+.container-wrap {
   //margin-top: 250px;
 }
 
-.container{
+.container {
   display: flex;
   gap: 10px;
+  justify-content: center;
 }
 
-.seat-wrap{
+.seat-wrap {
   padding: 20px;
   background-color: rgb(231, 219, 204);
   height: 240px;
@@ -142,7 +148,7 @@ export default {
   box-shadow: -2px 2px 5px rgba(0, 0, 0, 0.1);
 }
 
-.seat-table-wrap{
+.seat-table-wrap {
   position: absolute;
   padding: 10px;
   background-color: rgb(236, 234, 231);
@@ -151,89 +157,101 @@ export default {
   box-shadow: -2px 2px 5px rgba(0, 0, 0, 0.1);
 }
 
-.seat-table-label{
-
+.seat-table-label {
+  margin: 0;
 }
 
-.seat-table{
+.seat-table {
   width: 100px;
   box-sizing: border-box;
   background-color: rgb(189, 168, 141);
   display: flex;
   align-items: center;
   justify-content: center;
-  color:  rgb(78, 69, 57);
+  color: rgb(78, 69, 57);
   font-weight: bold;
   font-size: 18px;
   height: 50px;
   border-radius: 5px;
   box-shadow: -2px 2px 5px rgba(0, 0, 0, 0.1);
-  &:hover{
+
+  &:hover {
     background-color: rgb(255, 255, 255);
     color: rgb(189, 168, 141);
   }
-  &:focus{
+
+  &:focus {
     background-color: rgb(78, 69, 57);
     color: rgb(189, 168, 141);
   }
-  &:active{
+
+  &:active {
     background-color: rgb(78, 69, 57);
     color: rgb(189, 168, 141);
   }
 }
+
 .seat-table.selected {
   background-color: rgb(78, 69, 57);
   color: rgb(189, 168, 141);
 }
-.f3{
+
+.f3 {
   top: 20px;
   left: 20px;
 }
 
-.f4{
+.f4 {
   top: 120px;
   left: 20px;
 }
 
-.f5{
+.f5 {
   top: 40px;
   left: 150px;
   width: 270px;
-  .seat-table-fifth{
+
+  .seat-table-fifth {
     position: relative;
   }
+
   /* position: relative; */
-  .seat-table{
+
+  .seat-table {
     margin-bottom: 10px;
   }
 
-  .seat-table:nth-child(2), .seat-table:nth-child(4){
+  .seat-table:nth-child(2), .seat-table:nth-child(4) {
     margin-bottom: 0;
   }
-  .t3{
+
+  .t3 {
     left: 110px;
     position: absolute;
     top: 0;
   }
-  .t4{
+
+  .t4 {
     left: 110px;
     position: absolute;
     top: 60px;
   }
-  .t5{
+
+  .t5 {
     left: 220px;
     position: absolute;
     top: 0px;
     height: 110px;
     width: 50px;
-    &:disabled{
+
+    &:disabled {
       background-color: rgb(168, 161, 151);
       color: rgb(236, 234, 231);
     }
   }
 }
 
-.seating-status-wrap{
+.seating-status-wrap {
   background-color: rgb(242, 239, 234);
   padding: 8px;
   box-sizing: border-box;
@@ -247,7 +265,7 @@ export default {
   gap: 5px;
 }
 
-.seating-status{
+.seating-status {
   display: flex;
   gap: 10px;
   align-items: center;
@@ -256,27 +274,28 @@ export default {
   font-weight: bold;
   font-size: 12px;
 }
-.seating-status.available{
+
+.seating-status.available {
   color: rgb(78, 69, 57);
   background-color: rgb(189, 168, 141);
 }
 
-.seating-status.selected{
+.seating-status.selected {
   color: rgb(189, 168, 141);
   background-color: rgb(78, 69, 57);
 }
 
-.seating-status.unable{
+.seating-status.unable {
   color: rgb(236, 234, 231);
   background-color: rgb(168, 161, 151);
 }
 
-.seat-stat{
+.seat-stat {
   display: flex;
   /* position: relative; */
 }
 
-.seat-stat-announcement{
+.seat-stat-announcement {
   display: flex;
   align-items: center;
   justify-content: center;
@@ -289,8 +308,9 @@ export default {
   z-index: 1;
   position: absolute;
 }
-.seat-time-table{
-  z-index: 10;
+
+.seat-time-table {
+  z-index: 2;
   display: flex;
   /* width: 100%; */
   background-color: rgb(242, 239, 234);
@@ -302,43 +322,46 @@ export default {
   box-shadow: 2px 2px 5px rgba(0, 0, 0, 0.1);
 }
 
-.seat-time-table-wrap{
+.seat-time-table-wrap {
   display: flex;
   flex-direction: column;
   padding: 10px;
   justify-content: space-between;
-  p{
+
+  p {
     font-size: 18px;
     font-weight: bold;
     color: rgb(78, 69, 57);
   }
 }
 
-.time-select-wrap{
+.time-select-wrap {
   display: flex;
   height: 100%;
   flex-direction: column;
 }
 
-.time-selector{
+.time-selector {
   display: flex;
   flex-direction: column;
   gap: 5px;
   margin-top: 10px;
 }
 
-.time-select-btn{
+.time-select-btn {
   background-color: rgb(189, 168, 141);
   color: rgb(78, 69, 57);
   padding: 5px;
   border-radius: 5px;
   font-weight: bold;
   font-size: 16px;
-  &:hover{
+
+  &:hover {
     background-color: rgb(78, 69, 57);
     color: rgb(189, 168, 141);
   }
-  &:focus{
+
+  &:focus {
     background-color: rgb(78, 69, 57);
     color: rgb(189, 168, 141);
   }
