@@ -1,71 +1,86 @@
-<script>
+<script setup>
+import { ref, computed } from 'vue';
+import { useRouter } from 'vue-router';
+import MainHeader from '@/components/layout/MainHeader.vue';
+import SocialLoginBtn from '@/pages/User/component/SocialLoginBtn.vue';
+import SubmitBtn from '@/components/button/SubmitBtn.vue';
+import { useLoginStore } from '@/pages/User/stores/useLoginStore';
+import MainFooter from "@/components/layout/MainFooter.vue";
 
-import {defineComponent} from "vue";
-import MainHeader from "@/components/layout/MainHeader.vue";
-import SocialLoginBtn from "@/pages/User/component/SocialLoginBtn.vue";
-import SubmitBtn from "@/components/button/SubmitBtn.vue";
+const loginStore = useLoginStore();
+const router = useRouter();
+const userEmail = ref('');
+const password = ref('');
+const error = ref('');
 
-export default defineComponent({
-  name: "LoginPage",
-  data(){
-    return {
+const handleLogin = async () => {
+  error.value = ''; // Reset error message
+  try {
+    const success = await loginStore.login({ email: userEmail.value, password: password.value });
+    if (success) {
+      await router.push('/');
+    } else {
+      error.value = '로그인에 실패했습니다. 아이디와 비밀번호를 확인해주세요.';
     }
-  },
-  components: {
-    SocialLoginBtn,
-    MainHeader,
-    SubmitBtn
+  } catch (e) {
+    console.error(e);
+    error.value = '로그인에 실패했습니다. 아이디와 비밀번호를 확인해주세요.';
   }
-})
+};
+
+const isDisabled = computed(() => {
+  return !userEmail.value || !password.value;
+});
 </script>
 
 <template>
   <div class="body-container">
-    <MainHeader></MainHeader>
+    <MainHeader />
     <main>
-      <div class="body-container">
-        <div class="login-container">
-          <h2 class="login-title">Beyond SW 에 오신것을 환영합니다.</h2>
-          <div class="content-container">
-            <span class="social-login-title">SNS 로그인</span>
-            <SocialLoginBtn></SocialLoginBtn>
-            <div class="login-subtitle">
-              <hr>
-              <span> BOOT UP 아이디로 로그인 </span>
-              <hr>
-            </div>
-            <form class="login-form">
-              <div class="login-form-container">
-                <div class="id-wrapper">
-                  <label for="user-id">아이디</label>
-                  <input id="user-id" type="text" autocomplete="name" class="login-input" name="userId">
-                </div>
-                <div class="pw-wrapper">
-                  <label for="user-pw">비밀번호</label>
-                  <input id="user-pw" type="password" autocomplete="current-password" class="login-input" name="password">
-                </div>
-              </div>
-              <div class="button-container">
-                <SubmitBtn text="로그인"></SubmitBtn>
-              </div>
-              <p class="not-member-wrapper"><span>아직 회원이 아니신가요?</span>
-                <router-link class="signup-link" to="/signup-agree">회원가입</router-link>
-              </p>
-            </form>
+      <div class="login-container">
+        <h2 class="login-title">Beyond SW에 오신 것을 환영합니다.</h2>
+        <div class="content-container">
+          <span class="social-login-title">SNS 로그인</span>
+          <SocialLoginBtn />
+          <div class="login-subtitle">
+            <hr />
+            <span>BOOT UP 아이디로 로그인</span>
+            <hr />
           </div>
+          <form class="login-form" @submit.prevent="handleLogin">
+            <div class="login-form-container">
+              <div class="id-wrapper">
+                <label for="user-id">아이디</label>
+                <input id="user-id" type="text" autocomplete="name" class="login-input" v-model="userEmail" />
+              </div>
+              <div class="pw-wrapper">
+                <label for="user-pw">비밀번호</label>
+                <input id="user-pw" type="password" autocomplete="current-password" class="login-input" v-model="password" />
+              </div>
+            </div>
+            <div class="button-container">
+              <SubmitBtn :handle-submit="handleLogin" :is-disabled="isDisabled" text="로그인"></SubmitBtn>
+            </div>
+            <p class="not-member-wrapper">
+              <span>아직 회원이 아니신가요?</span>
+              <router-link class="signup-link" to="/user/signup/agree">회원가입</router-link>
+            </p>
+            <p v-if="error" class="error-message">{{ error }}</p>
+          </form>
         </div>
       </div>
     </main>
+    <MainFooter />
   </div>
 </template>
 
 <style>
-.body-container{
+.body-container {
   width: 100%;
-  //max-width: 800px;
 }
+
 .login-container {
-  max-width: 500px;
+  max-width: 800px;
   width: 100%;
 }
 
@@ -79,8 +94,9 @@ export default defineComponent({
   font-weight: 700;
   color: #E06139;
   display: flex;
-  hr{
-    width: 150px;
+
+  hr {
+    width: 300px;
     height: 2px;
     background-color: #E06139;
     border: 0;
@@ -99,14 +115,13 @@ export default defineComponent({
   gap: 20px;
 }
 
-.login-input{
+.login-input {
   border-radius: 5px;
   border: 1px solid hsla(220, 9%, 46%, .3);
   box-shadow: 0 1px 2px 0 rgba(0, 0, 0, .05);
   box-sizing: border-box;
   line-height: 1.5rem;
   width: 100%;
-  border-radius: 5px;
   margin-top: 10px;
 }
 
@@ -115,7 +130,7 @@ export default defineComponent({
 }
 
 .not-member-wrapper {
-  gap : 10px;
+  gap: 10px;
   display: flex;
   justify-content: center;
 }
