@@ -7,7 +7,7 @@
 
 <script>
 export default {
-  name: 'SocialButtonComponent',
+  name: 'SocialLoginBtn',
   props: {
     msg: String
   }
@@ -18,6 +18,7 @@ export default {
 <style scoped>
 .social-button-wrapper {
   margin-top: 5px;
+  width: 100%;
 }
 
 .social-button {
