@@ -15,20 +15,21 @@ const router = createRouter({
                 { path: 'login', component: () => import('../pages/User/LoginPage.vue') },
                 { path: 'signup/agree', component: () => import('../pages/User/SignupAgree.vue') },
                 { path: 'signup/info', component: () => import('../pages/User/SignupPage.vue') },
-                { path: 'signup/complete', component: () => import('../pages/User/SignupComplete.vue') },
+                // { path: 'signup/complete', component: () => import('../pages/User/SignupComplete.vue') },
+                { path: 'signup/success', component: () => import('../pages/User/SignupSuccessPage.vue') }
             ]
         },
 
         // Community
         { path: '/community', component: () => import('../pages/Community/CommunityBoardListPage.vue') },
-        { path: '/community/free', component: () => import('../pages/Community/FreeBoard/FreeListPage.vue') },
-        { path: '/community/free/detail/:id', component: () => import('../pages/Community/FreeBoard/FreeDetailPage.vue') },
-        { path: '/community/free/write', component: () => import('../pages/Community/FreeBoard/FreeWritePage.vue') },
-        { path: '/community/free/edit/:id', component: () => import('../pages/Community/FreeBoard/FreeEditPage.vue') },
-        { path: '/community/open', component: () => import('../pages/Community/OpenBoard/OpenListPage.vue') },
-        { path: '/community/open/detail/:id', component: () => import('../pages/Community/OpenBoard/OpenDetailPage.vue') },
-        { path: '/community/open/write', component: () => import('../pages/Community/OpenBoard/OpenWritePage.vue') },
-        { path: '/community/open/edit/:id', component: () => import('../pages/Community/OpenBoard/OpenEditPage.vue') },
+        { path: '/free', component: () => import('../pages/Community/FreeBoard/FreeListPage.vue') },
+        { path: '/free/detail/:id', component: () => import('../pages/Community/FreeBoard/FreeDetailPage.vue') },
+        { path: '/free/write', component: () => import('../pages/Community/FreeBoard/FreeWritePage.vue') },
+        { path: '/free/edit/:id', component: () => import('../pages/Community/FreeBoard/FreeEditPage.vue') },
+        { path: '/open', component: () => import('../pages/Community/OpenBoard/OpenListPage.vue') },
+        { path: '/open/detail/:id', component: () => import('../pages/Community/OpenBoard/OpenDetailPage.vue') },
+        { path: '/open/write', component: () => import('../pages/Community/OpenBoard/OpenWritePage.vue') },
+        { path: '/open/edit/:id', component: () => import('../pages/Community/OpenBoard/OpenEditPage.vue') },
 
         // Notice Board
         { path: '/notice', component: () => import('../pages/Notice/NoticeListPage.vue') },
@@ -44,7 +45,7 @@ const router = createRouter({
 
         // Market Board
         { path: '/market', component: () => import('../pages/Market/Board/MarketListPage.vue') },
-        { path: '/market/detail/:id', component: () => import('../pages/Market/Board/MarketDetailPage.vue') },
+        { path: '/market/detail/:idx', component: () => import('../pages/Market/Board/MarketDetailPage.vue') },
         { path: '/market/write', component: () => import('../pages/Market/Board/MarketWritePage.vue') },
         // Market Chat
         { path: '/market/chat', component: () => import('../pages/Market/Chat/MarketChatPage.vue') },
