@@ -1,7 +1,14 @@
-const { defineConfig } = require('@vue/cli-service')
+const { defineConfig } = require("@vue/cli-service");
 module.exports = defineConfig({
-    transpileDependencies: true,
     devServer: {
-        port: 3001
-    }
-})
+        port: 3000,
+        proxy: {
+            "/api": {
+                target: "http://localhost:8080",
+                changeOrigin: true,
+                pathRewrite: { "^/api": "" },
+            },
+        },
+    },
+    transpileDependencies: true,
+});
